back/requirements.txt
aiofiles>=23.2.0
aiohttp>=3.10.11
APScheduler>=3.10.4
bcrypt
ccxt>=4.4.61
certifi>=2025.1.31
charset-normalizer>=3.4.1
click>=8.1.8
colorama>=0.4.6
cryptocompare
cryptography>=44.0.1
dash>=2.18.2
dash-bootstrap-components>=1.7.1
dash-core-components>=2.0.0
dash-html-components>=2.0.0
dash-table>=5.0.0
Flask>=2.3.3
Flask-CORS>=4.0.0
Flask-Login>=0.6.3
Flask-JWT-Extended>=4.5.3
Jinja2>=3.1.2
MarkupSafe>=2.1.3
numpy>=1.24.3
pandas>=2.1.1
psycopg2-binary>=2.9.7
pyTelegramBotAPI>=4.14.0
python-binance>=1.0.19
python-dotenv>=1.0.0
pytz>=2023.3
schedule>=1.2.0
redis>=5.0.1
requests>=2.31.0
supabase>=2.0.0
ta>=0.10.2
tabulate
tqdm
websocket-client>=1.8.0
Werkzeug>=2.3.7
waitress>=2.1.2
psutil>=5.9.0